server/requirements.txt
flask
marshmallow
marshmallow-sqlalchemy
flask_sqlalchemy
flask_cors
flask-restx
python-dotenv